chore: move format function to helper
